Solution to Manage Book Library

+++ Manage Book Library@@67877d61dc107486ad5c274c/code.c
@@ -58,7 +58,7 @@ int main(int argc, char *const argv[])
     printf("Books above price %0.2f\n", (float)price);
     for (int i = 0; i < n; i++) {
         if (students[i].price >= price) {
-            printf("Title: %s, Author: %s, Price: %0.2f\n", students[i].title, students[i].author, students[i].price);
+            printf("Title: %s, Author: %s, Price: %0.2f\n", students[i].title, students[i].author, (float)students[i].price);
         }
         free(students[i].title);
         free(students[i].author);
